Update the 2 notification settings screens by moving some pref to the root notification screen
