+++ .idea/dictionaries/bmarty.xml
@@ -19,6 +19,7 @@
       <w>msisdn</w>
       <w>pbkdf</w>
       <w>pkcs</w>
+      <w>riotx</w>
       <w>signin</w>
       <w>signout</w>
       <w>signup</w>
