+++ vector/src/main/java/im/vector/riotx/features/settings/VectorSettingsNotificationPreferenceFragment.kt
@@ -16,6 +16,11 @@
 
 package im.vector.riotx.features.settings
 
+import android.app.Activity
+import android.content.Intent
+import android.media.RingtoneManager
+import android.net.Uri
+import android.os.Parcelable
 import android.widget.Toast
 import androidx.preference.Preference
 import androidx.preference.SwitchPreference
@@ -24,8 +29,10 @@ import im.vector.matrix.android.api.pushrules.RuleIds
 import im.vector.matrix.android.api.pushrules.RuleKind
 import im.vector.riotx.R
 import im.vector.riotx.core.di.ActiveSessionHolder
+import im.vector.riotx.core.preference.VectorPreference
 import im.vector.riotx.core.preference.VectorSwitchPreference
 import im.vector.riotx.core.pushers.PushersManager
+import im.vector.riotx.features.notifications.NotificationUtils
 import im.vector.riotx.push.fcm.FcmHelper
 import javax.inject.Inject
 
@@ -54,7 +61,81 @@ class VectorSettingsNotificationPreferenceFragment @Inject constructor(
             val areNotifEnabledAtAccountLevel = !mRuleMaster.enabled
             (pref as SwitchPreference).isChecked = areNotifEnabledAtAccountLevel
         }
+
+        handleSystemPreference()
     }
+
+    private fun handleSystemPreference() {
+        val callNotificationsSystemOptions = findPreference<VectorPreference>(VectorPreferences.SETTINGS_SYSTEM_CALL_NOTIFICATION_PREFERENCE_KEY)!!
+        if (NotificationUtils.supportNotificationChannels()) {
+            callNotificationsSystemOptions.onPreferenceClickListener = Preference.OnPreferenceClickListener {
+                NotificationUtils.openSystemSettingsForCallCategory(this)
+                false
+            }
+        } else {
+            callNotificationsSystemOptions.isVisible = false
+        }
+
+        val noisyNotificationsSystemOptions = findPreference<VectorPreference>(VectorPreferences.SETTINGS_SYSTEM_NOISY_NOTIFICATION_PREFERENCE_KEY)!!
+        if (NotificationUtils.supportNotificationChannels()) {
+            noisyNotificationsSystemOptions.onPreferenceClickListener = Preference.OnPreferenceClickListener {
+                NotificationUtils.openSystemSettingsForNoisyCategory(this)
+                false
+            }
+        } else {
+            noisyNotificationsSystemOptions.isVisible = false
+        }
+
+        val silentNotificationsSystemOptions = findPreference<VectorPreference>(VectorPreferences.SETTINGS_SYSTEM_SILENT_NOTIFICATION_PREFERENCE_KEY)!!
+        if (NotificationUtils.supportNotificationChannels()) {
+            silentNotificationsSystemOptions.onPreferenceClickListener = Preference.OnPreferenceClickListener {
+                NotificationUtils.openSystemSettingsForSilentCategory(this)
+                false
+            }
+        } else {
+            silentNotificationsSystemOptions.isVisible = false
+        }
+
+        // Ringtone
+        val ringtonePreference = findPreference<VectorPreference>(VectorPreferences.SETTINGS_NOTIFICATION_RINGTONE_SELECTION_PREFERENCE_KEY)!!
+
+        if (NotificationUtils.supportNotificationChannels()) {
+            ringtonePreference.isVisible = false
+        } else {
+            ringtonePreference.summary = vectorPreferences.getNotificationRingToneName()
+            ringtonePreference.onPreferenceClickListener = Preference.OnPreferenceClickListener {
+                val intent = Intent(RingtoneManager.ACTION_RINGTONE_PICKER)
+                intent.putExtra(RingtoneManager.EXTRA_RINGTONE_TYPE, RingtoneManager.TYPE_NOTIFICATION)
+
+                if (null != vectorPreferences.getNotificationRingTone()) {
+                    intent.putExtra(RingtoneManager.EXTRA_RINGTONE_EXISTING_URI, vectorPreferences.getNotificationRingTone())
+                }
+
+                startActivityForResult(intent, REQUEST_NOTIFICATION_RINGTONE)
+                false
+            }
+        }
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (resultCode == Activity.RESULT_OK) {
+            when (requestCode) {
+                REQUEST_NOTIFICATION_RINGTONE -> {
+                    vectorPreferences.setNotificationRingTone(data?.getParcelableExtra<Parcelable>(RingtoneManager.EXTRA_RINGTONE_PICKED_URI) as Uri?)
+
+                    // test if the selected ring tone can be played
+                    val notificationRingToneName = vectorPreferences.getNotificationRingToneName()
+                    if (null != notificationRingToneName) {
+                        vectorPreferences.setNotificationRingTone(vectorPreferences.getNotificationRingTone())
+                        findPreference<VectorPreference>(VectorPreferences.SETTINGS_NOTIFICATION_RINGTONE_SELECTION_PREFERENCE_KEY)!!
+                                .summary = notificationRingToneName
+                    }
+                }
+            }
+        }
+    }
+
 
     override fun onResume() {
         super.onResume()
@@ -128,5 +209,9 @@ class VectorSettingsNotificationPreferenceFragment @Inject constructor(
                                 }
                             })
                 }
+    }
+
+    companion object {
+        private const val REQUEST_NOTIFICATION_RINGTONE = 888
     }
 }
